Fixed bug in control module instantiation, control works now
The problem with control not working was that a new signal called
"clk" had been defined and was used as the input clock signal
for the control module, but "clk" was not connected to anything.
Connected input clock to adcClk instead.

We definitely need some diagnostic capabilities for optimising the lock