Added/tested FIFO routing.  PID VHDL array has timing issues
So it looks like when signals are defined as arrays that Vivado tries
to keep the signals together, and this creates routing issues.
When the pid register signals were all separate, routing didn't
take too long and the design met timing constraints.  Now that they
are all arrays of arrays of arrays, timing is not met.  This needs
testing