Added code + built bitstream for phase lock at second harmonic
Significant modification to design to allow for phase measurement
and locking.  Added a module PhaseControl which comprises a phase
measurement module, a phase unwrapping module, and a PID controller.
Tested this module in simulation and it seems to work.

I have assumed for the moment that the phase lock actuator
is the 4th PWM output.  This should get changed in the future
to allow switching between that output and the aux DAC.

Added extra FIFOs for storing phase data.